do an example for the entire v2 impacted files endpoint
